submit_report.sh, kdump.etc: Add setting to disable log submission
Users may want to not submit logs to Valve servers, either for debug
purposes or just due to their preference. This patch adds a setting
for that, exposed in /etc/default/kdump . Information about this
tuning was added to README as well.

Finally, this commit also improves journal output when we bail-out
without submitting the logs to Valve servers, showing a friendly
message pointing to the locally saved file.

Signed-off-by: Guilherme G. Piccoli <gpiccoli@igalia.com>