module-setup.sh: Bail-out if makedumpfile is not available
If we don't have makedumpfile, it doesn't make sense to construct
the kdump initrd and let it be loaded; it's going to fail in the
kdump dmesg collection, during a panic event, with no clear traces
for users to diagnose the issue.

So, let's bail-out if we don't have makedumpfile, forcing the
kdump load to fail instead, which is clearly warned in journalctl.
Also, change the approach for the kdump.conf file as well, in
order to fail creating the initrd if any of the files are missing.

While at it, fix a trailing space in the module-setup.sh file.

Signed-off-by: Guilherme G. Piccoli <gpiccoli@igalia.com>