kdump-load.sh: Attempt creating initrd before kdump load, if we don't have one
There might be a (rare) case of missing initrd when loading a kdump.
It's rare mainly for 2 reasons:

(a) Pstore is the default log collection mechanism, kdump should only
be used as a fallback;

(b) When the package is installed, initrd is created for the
running kernel.

But imagine the user installs a new kernel with no Deck image upgrade;
this would cause the issue of a missing initrd if/when kdump is loaded.

We hereby fix it by attempting to create the initrd before kdump load,
in case it doesn't exist.

Signed-off-by: Guilherme G. Piccoli <gpiccoli@igalia.com>