kdump-load: Fix the initrd clean-up
This is not really related to the upstreaming effort, but many
changes happened so it's a lot easier to just add this patch on
top of it.

The issues are two: first, if the kdump directory is empty, the
"find" tool complains, which is harmless but an unnecessary log
pollution. But more important, the second issue is related to
writing the kernel versions file, which was unreliable - we now
fix that by being more explicit in the usage of stdout redirection.

Signed-off-by: Guilherme G. Piccoli <gpiccoli@igalia.com>