20-kdump-steamos.conf: don't panic on RCU stalls
I found a small RCU stall in v6.1-rc3 that is quite insignificant,
but panick'ed the system. I feel we're in a risk to have it
enabled by default, seems better to stay in the "safe" side
of the trade-off and eventually lose some report than to panic
in meaningless stall and affect user experience.

Signed-off-by: Guilherme G. Piccoli <gpiccoli@igalia.com>