kdump-load: Improve vmlinux binary path obtaining
Currently we just look the command-line for the systemd BOOT_IMAGE
entry and try to load the vmlinux as described there.

But in the field we noticed we have 2 patterns: either this cmdline
entry is complete/correct and maps to the vmlinux binary, OR it shows
the name of the file but not the full path (Arch case). So, let's
hereby attempt prepending /boot in case the validation for the first
case fails - if both ways don't work, we just fail kdump for now.

Signed-off-by: Guilherme G. Piccoli <gpiccoli@igalia.com>