module-setup.sh: Include nvme (if module) in the dracut initramfs image
On newer Arch kernels, the nvme module started to be built as module;
previously, it was built-in. Some discussions about this decision were
found by @cristicc on issue jupiter/#718.

In the end, there is not strong reason for any approach, so seems
that Valve is sticking with te module approach, following Arch (and
many distros). With that said, we need to include the nvme module in
the dracut initramfs to prevent potential boot stalls during the kdump
panic kernel load, due to lack of rootfs module in the initrd.

Signed-off-by: Guilherme G. Piccoli <gpiccoli@igalia.com>