README: Update documentation
Documentation now mentions the new name of the tool (kdumpst)
and the support of both initcpio/dracut; also updated the pstore
instructions - users might need to reserve some memory for that,
not all HWs are like the Steam Deck, which contains some pre-reserved
region of memory. Improved the TODO section a bit too.

Signed-off-by: Guilherme G. Piccoli <gpiccoli@igalia.com>